requirements.txt
python-dotenv==1.0.0
pyaudio==0.2.14
pvporcupine==3.0.0
vosk==0.3.45
spacy==3.7.2
python-dateutil==2.8.2
ollama==0.1.6
sqlalchemy==2.0.23
apscheduler==3.10.4
RPi.GPIO==0.7.1
colorlog==6.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
piper-tts==1.2.0
onnxruntime==1.16.3
numpy==1.24.3
